Fixed loading without using Ext.ux.data.PagingStore
Store.load paging wins' very easy for local load but that's just an
array. Don't really want to waste time on fixing this to work with
Ext.ux.data.PagingStore because we are not gonna be loading data from
local very much anymore. Eventually the store code will be pulling data
from a proxy.